[FIRRTL] Fix Windows build size problem
Fix some code that was not portable to Windows due to "0ul" being 4-byte
on Windows and 8-byte on everything else.

+++ lib/Dialect/FIRRTL/Transforms/ExpandWhens.cpp
@@ -223,7 +223,7 @@ public:
           }
         })
         .template Case<FVectorType>([&](FVectorType vector) {
-          for (auto i : llvm::seq(0ul, vector.getNumElements())) {
+          for (auto i : llvm::seq((size_t)0, vector.getNumElements())) {
             auto subindex =
                 builder.create<SubindexOp>(value.getLoc(), value, i);
             foreachSubelement(builder, subindex, fn);
